Fixed issues from PR review, also reworked the way how to choose storage type in stress-test module
