+++ .idea/codeStyleSettings.xml
@@ -3,6 +3,7 @@
   <component name="ProjectCodeStyleSettingsManager">
     <option name="PER_PROJECT_SETTINGS">
       <value>
+        <option name="GENERATE_FINAL_LOCALS" value="true" />
         <option name="CLASS_COUNT_TO_USE_IMPORT_ON_DEMAND" value="1000" />
         <option name="NAMES_COUNT_TO_USE_IMPORT_ON_DEMAND" value="1000" />
         <option name="RIGHT_MARGIN" value="100" />
