+++ ext.gradle
@@ -43,6 +43,6 @@ project.ext {
     mysqlDriverVersion = '6.0.6'
 
     // Test scope third party dependencies' versions
-    jUnitPlatformVersion = '1.0.0-RC3'
+    jUnitPlatformVersion = '1.0.0'
     jUnitVersion = '5.0.0'
 }
