+++ stress-test/src/test/java/io/spine/test/integration/ToDoListTest.java
@@ -35,7 +35,7 @@ import io.spine.examples.todolist.c.commands.UpdateTaskDueDate;
 import io.spine.examples.todolist.c.commands.UpdateTaskPriority;
 import io.spine.examples.todolist.client.TodoClient;
 import io.spine.examples.todolist.q.projection.TaskItem;
-import io.spine.test.integration.BaseIT;
+import io.spine.test.BaseStressTest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -55,7 +55,10 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-public class ToDoListIT extends BaseIT {
+/**
+ * @author Dmitry Ganzha
+ */
+public class ToDoListTest extends BaseStressTest {
     private TodoClient client;
 
     @BeforeEach
