+++ stress-test/src/test/resources/jdbc-storage.properties
@@ -20,7 +20,6 @@
 
 # Default configuration properties for the local JDBC storage.
 
-db.prefix=jdbc:mysql:
 db.name=TodoListDB
 db.username=root
 db.password=root
